♻️ JE-10-refactor: passwordInput 분리
